# packages/api-client/openapi/onboarding.yaml
openapi: 3.1.0
info:
  title: Royal Palace Onboarding API
  version: 1.0.0
  description: Authenticated applicant onboarding and bounded support/administrator review through the same-origin BFF.
servers:
  - url: /
security:
  - sessionCookie: []
paths:
  /api/applications:
    get:
      operationId: listOwnApplications
      parameters: &listParameters
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/ApplicationKind"
        - $ref: "#/components/parameters/ApplicationStatus"
      responses:
        "200": { $ref: "#/components/responses/ApplicationList" }
        default: { $ref: "#/components/responses/Error" }
  /api/applications/patients:
    post:
      operationId: createPatientApplication
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      requestBody: { $ref: "#/components/requestBodies/PatientData" }
      responses:
        "201": { $ref: "#/components/responses/ApplicationDetail" }
        default: { $ref: "#/components/responses/Error" }
  /api/applications/patients/{applicationId}:
    parameters: [{ $ref: "#/components/parameters/ApplicationId" }]
    get:
      operationId: getPatientApplication
      responses:
        "200": { $ref: "#/components/responses/ApplicationDetail" }
        default: { $ref: "#/components/responses/Error" }
    patch:
      operationId: updatePatientApplication
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/VersionedCommand"
                - type: object
                  required: [values]
                  properties: { values: { $ref: "#/components/schemas/PatientApplicationData" } }
      responses:
        "200": { $ref: "#/components/responses/ApplicationDetail" }
        default: { $ref: "#/components/responses/Error" }
  /api/applications/organizations:
    post:
      operationId: createOrganizationApplication
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      requestBody: { $ref: "#/components/requestBodies/OrganizationData" }
      responses:
        "201": { $ref: "#/components/responses/ApplicationDetail" }
        default: { $ref: "#/components/responses/Error" }
  /api/applications/organizations/{applicationId}:
    parameters: [{ $ref: "#/components/parameters/ApplicationId" }]
    get:
      operationId: getOrganizationApplication
      responses:
        "200": { $ref: "#/components/responses/ApplicationDetail" }
        default: { $ref: "#/components/responses/Error" }
    patch:
      operationId: updateOrganizationApplication
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/VersionedCommand"
                - type: object
                  required: [values]
                  properties:
                    { values: { $ref: "#/components/schemas/OrganizationApplicationData" } }
      responses:
        "200": { $ref: "#/components/responses/ApplicationDetail" }
        default: { $ref: "#/components/responses/Error" }
  /api/applications/practitioners:
    post:
      operationId: createPractitionerApplication
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      requestBody: { $ref: "#/components/requestBodies/PractitionerData" }
      responses:
        "201": { $ref: "#/components/responses/ApplicationDetail" }
        default: { $ref: "#/components/responses/Error" }
  /api/applications/practitioners/{applicationId}:
    parameters: [{ $ref: "#/components/parameters/ApplicationId" }]
    get:
      operationId: getPractitionerApplication
      responses:
        "200": { $ref: "#/components/responses/ApplicationDetail" }
        default: { $ref: "#/components/responses/Error" }
    patch:
      operationId: updatePractitionerApplication
      parameters: [{ $ref: "#/components/parameters/Csrf" }]
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/VersionedCommand"
                - type: object
                  required: [values]
                  properties:
                    { values: { $ref: "#/components/schemas/PractitionerApplicationData" } }
      responses:
        "200": { $ref: "#/components/responses/ApplicationDetail" }
        default: { $ref: "#/components/responses/Error" }
  /api/applications/{applicationId}/documents/upload-intents:
    post:
      operationId: reserveApplicationDocument
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
        - $ref: "#/components/parameters/Csrf"
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DocumentIntentRequest" }
      responses:
        "201":
          description: Short-lived, single-write quarantine upload intent.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DocumentIntentResponse" }
        default: { $ref: "#/components/responses/Error" }
  /api/applications/{applicationId}/documents/{documentId}/complete:
    post:
      operationId: completeApplicationDocumentUpload
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
        - $ref: "#/components/parameters/DocumentId"
        - $ref: "#/components/parameters/Csrf"
      responses:
        "200": { $ref: "#/components/responses/DocumentCompletion" }
        default: { $ref: "#/components/responses/Error" }
  /api/applications/{applicationId}/documents/{documentId}/download:
    get:
      operationId: downloadOwnApplicationDocument
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200": { $ref: "#/components/responses/DocumentDownload" }
        default: { $ref: "#/components/responses/Error" }
  /api/applications/{applicationId}/submit:
    post:
      operationId: submitApplication
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
        - $ref: "#/components/parameters/Csrf"
      requestBody: { $ref: "#/components/requestBodies/VersionedCommand" }
      responses:
        "200": { $ref: "#/components/responses/ApplicationDetail" }
        default: { $ref: "#/components/responses/Error" }
  /api/applications/{applicationId}/withdraw:
    post:
      operationId: withdrawApplication
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
        - $ref: "#/components/parameters/Csrf"
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/VersionedCommand"
                - type: object
                  properties: { note: { type: string, maxLength: 2000 } }
      responses:
        "200": { $ref: "#/components/responses/ApplicationDetail" }
        default: { $ref: "#/components/responses/Error" }
  /api/support/applications:
    get:
      operationId: listSupportApplications
      parameters: *listParameters
      responses:
        "200": { $ref: "#/components/responses/ApplicationList" }
        default: { $ref: "#/components/responses/Error" }
  /api/support/applications/{applicationId}:
    get:
      operationId: getSupportApplication
      parameters: [{ $ref: "#/components/parameters/ApplicationId" }]
      responses:
        "200": { $ref: "#/components/responses/ApplicationDetail" }
        default: { $ref: "#/components/responses/Error" }
  /api/support/applications/{applicationId}/documents/{documentId}/download:
    get:
      operationId: downloadSupportApplicationDocument
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200": { $ref: "#/components/responses/DocumentDownload" }
        default: { $ref: "#/components/responses/Error" }
  /api/support/applications/{applicationId}/start-review:
    post:
      operationId: startApplicationReview
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
        - $ref: "#/components/parameters/Csrf"
      requestBody: { $ref: "#/components/requestBodies/VersionedCommand" }
      responses:
        "200": { $ref: "#/components/responses/ApplicationDetail" }
        default: { $ref: "#/components/responses/Error" }
  /api/admin/applications:
    get:
      operationId: listAdminApplications
      parameters: *listParameters
      responses:
        "200": { $ref: "#/components/responses/ApplicationList" }
        default: { $ref: "#/components/responses/Error" }
  /api/admin/applications/{applicationId}:
    get:
      operationId: getAdminApplication
      parameters: [{ $ref: "#/components/parameters/ApplicationId" }]
      responses:
        "200": { $ref: "#/components/responses/ApplicationDetail" }
        default: { $ref: "#/components/responses/Error" }
  /api/admin/applications/{applicationId}/documents/{documentId}/download:
    get:
      operationId: downloadAdminApplicationDocument
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200": { $ref: "#/components/responses/DocumentDownload" }
        default: { $ref: "#/components/responses/Error" }
  /api/admin/applications/{applicationId}/approve:
    post:
      operationId: approveApplication
      parameters: &decisionParameters
        - $ref: "#/components/parameters/ApplicationId"
        - $ref: "#/components/parameters/Csrf"
      requestBody: { $ref: "#/components/requestBodies/Decision" }
      responses: &decisionResponses
        "200": { $ref: "#/components/responses/ApplicationDetail" }
        default: { $ref: "#/components/responses/Error" }
  /api/admin/applications/{applicationId}/reject:
    post:
      operationId: rejectApplication
      parameters: *decisionParameters
      requestBody: { $ref: "#/components/requestBodies/Decision" }
      responses: *decisionResponses
  /api/admin/applications/{applicationId}/request-information:
    post:
      operationId: requestApplicationInformation
      parameters: *decisionParameters
      requestBody: { $ref: "#/components/requestBodies/Decision" }
      responses: *decisionResponses
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-session
  parameters:
    ApplicationId:
      in: path
      name: applicationId
      required: true
      schema: { type: string, format: uuid }
    DocumentId:
      in: path
      name: documentId
      required: true
      schema: { type: string, format: uuid }
    ApplicationKind:
      in: query
      name: kind
      schema: { $ref: "#/components/schemas/ApplicationKind" }
    ApplicationStatus:
      in: query
      name: status
      schema: { $ref: "#/components/schemas/ApplicationStatus" }
...
    Csrf:
      in: header
      name: x-rp-csrf-token
      required: true
      schema: { type: string, minLength: 1 }
    Cursor:
      in: query
      name: cursor
      schema: { type: string, maxLength: 2048 }
    Limit:
      in: query
      name: limit
      schema: { type: integer, minimum: 1, maximum: 50 }
  requestBodies:
    PatientData:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/PatientApplicationCreate" }
    OrganizationData:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/OrganizationApplicationCreate" }
    PractitionerData:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/PractitionerApplicationData" }
    VersionedCommand:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/VersionedCommand" }
    Decision:
      required: true
      content:
        application/json:
          schema: { $ref: "#/components/schemas/DecisionCommand" }
...
  responses:
    ApplicationDetail:
      description: Onboarding application detail.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApplicationDetail" }
    ApplicationList:
      description: Bounded onboarding application page.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApplicationListResponse" }
    DocumentCompletion:
      description: Upload verified and queued for malware scanning.
      content:
        application/json:
          schema:
            type: object
            required: [status]
            properties:
              status: { type: string, const: QUARANTINED }
    DocumentDownload:
      description: Short-lived download for an authorized clean document.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/DocumentDownloadResponse" }
    Error:
      description: Safe API error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
    ApplicationKind:
      type: string
      enum: [PATIENT, ORGANIZATION, PRACTITIONER]
    ApplicationStatus:
      type: string
      enum:
        [DRAFT, SUBMITTED, UNDER_REVIEW, MORE_INFORMATION_REQUIRED, APPROVED, REJECTED, WITHDRAWN]
    OrganizationType:
      type: string
      enum: [HOSPITAL, PHARMACY, LABORATORY]
    VersionedCommand:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      properties: { expectedVersion: { type: integer, minimum: 1 } }
...
    PatientApplicationData:
      type: object
      additionalProperties: false
      required: [countryCode, dateOfBirth, familyName, givenName, phoneE164, preferredLanguage]
      properties:
        countryCode: { type: [string, "null"] }
        dateOfBirth: { type: [string, "null"], format: date }
        familyName: { type: string, maxLength: 100 }
        givenName: { type: string, maxLength: 100 }
        phoneE164: { type: [string, "null"] }
        preferredLanguage: { type: [string, "null"] }
...
    OrganizationApplicationData:
      type: object
      additionalProperties: false
      required:
        [
          addressLine1,
          addressLine2,
          administrativeArea,
          contactEmail,
          contactName,
          contactPhoneE164,
          countryCode,
          displayName,
          jurisdictionCode,
          legalName,
          locality,
          organizationType,
          postalCode,
          registrationAuthority,
          registrationNumber,
          serviceIds,
        ]
      properties:
        addressLine1: { type: [string, "null"] }
        addressLine2: { type: [string, "null"] }
        administrativeArea: { type: [string, "null"] }
        contactEmail: { type: string, format: email }
        contactName: { type: string }
        contactPhoneE164: { type: [string, "null"] }
        countryCode: { type: string }
        displayName: { type: string }
        jurisdictionCode: { type: string }
        legalName: { type: string }
        locality: { type: [string, "null"] }
        organizationType: { $ref: "#/components/schemas/OrganizationType" }
        postalCode: { type: [string, "null"] }
        registrationAuthority: { type: string }
        registrationNumber: { type: string }
        serviceIds: { type: array, maxItems: 250, items: { type: string, format: uuid } }
...
    PractitionerApplicationData:
      type: object
      additionalProperties: false
      required:
        [
          biography,
          credentialType,
          facilityName,
          familyName,
          givenName,
          honorific,
          jurisdictionCode,
          professionIds,
          registrationAuthority,
          registrationNumber,
          selectedOrganizationId,
          specialtyIds,
        ]
      properties:
        biography: { type: [string, "null"] }
        credentialType: { type: string }
        facilityName: { type: [string, "null"] }
        familyName: { type: string }
        givenName: { type: string }
        honorific: { type: [string, "null"] }
        jurisdictionCode: { type: string }
        professionIds: { type: array, maxItems: 50, items: { type: string, format: uuid } }
        registrationAuthority: { type: string }
        registrationNumber: { type: string }
        selectedOrganizationId: { type: [string, "null"], format: uuid }
        specialtyIds: { type: array, maxItems: 100, items: { type: string, format: uuid } }
...
    ApplicationListResponse:
      type: object
      required: [data, pageInfo]
      properties:
        data: { type: array, items: { $ref: "#/components/schemas/ApplicationSummary" } }
        pageInfo: { $ref: "#/components/schemas/PageInfo" }
...
    DocumentIntentRequest:
      type: object
      additionalProperties: false
      required: [declaredContentType, declaredSha256, declaredSizeBytes, originalFilename, purpose]
      properties:
        declaredContentType: { type: string }
        declaredSha256: { type: string, pattern: "^[0-9a-f]{64}$" }
        declaredSizeBytes: { type: integer, minimum: 1, maximum: 26214400 }
        originalFilename: { type: string }
        purpose: { type: string }
...
    DocumentIntentResponse:
      type: object
      required: [document, upload]
      properties:
        document: { $ref: "#/components/schemas/DocumentMetadata" }
        upload:
          type: object
          required: [expiresAt, headers, method, url]
          properties:
            expiresAt: { type: string, format: date-time }
            headers: { type: object, additionalProperties: { type: string } }
            method: { type: string, const: PUT }
            url: { type: string, format: uri }
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error: { type: string }
        message: { type: string }
...